requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
channels==4.0.0
channels-redis==4.2.0
cssselect2==0.7.0
Django==5.1.6
django-multitenant==4.1.1
fonttools==4.54.1
gunicorn==23.0.0
html5lib==1.1
idna==3.10
matplotlib==3.8.3
mysql-connector-python==9.1.0
numpy==1.26.4
packaging==24.1
pandas==2.2.3
pillow==10.4.0
plotly==5.24.1
psycopg2-binary>=2.9.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tenacity==9.0.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2024.2
urllib3==2.2.3
weasyprint==63.0
webencodings==0.5.1
whitenoise==6.8.2
zopfli==0.2.3.post1
django-encrypted-model-fields==0.6.5
django-ckeditor-5==0.2.17